feat: add flag to write, print stats at the end

+++ src/optimizer.rs
@@ -1,18 +1,27 @@
 use crate::region_loader::region::Region;
-use indicatif::ProgressBar;
 use std::path::PathBuf;
 
-pub fn optimize_region_file(region_file_path: &PathBuf, pb: &ProgressBar) -> std::io::Result<()> {
-    let region_file_name = region_file_path
-        .file_stem()
-        .map(|os_str_val| os_str_val.to_string_lossy().into_owned())
-        .unwrap_or("Unk".to_string());
+#[derive(Default, Debug, Clone)]
+pub struct OptimizeResult {
+    pub total_chunks: usize,
+    pub deleted_chunks: usize,
+    pub deleted_regions: usize,
+}
+
+pub fn optimize_region_file(
+    region_file_path: &PathBuf,
+    write: bool,
+) -> std::io::Result<OptimizeResult> {
+    let mut result = OptimizeResult::default();
 
     match Region::from_file_name(region_file_path) {
         Ok(mut region) => {
-            // Read the first chunk
+            let chunks = region.get_chunks();
+
             let mut chunks_to_delete = Vec::new();
-            for chunk in region.get_chunks() {
+            result.total_chunks += chunks.len();
+
+            for chunk in chunks {
                 if chunk.should_delete() {
                     chunks_to_delete.push(chunk.clone());
                 }
@@ -22,28 +31,30 @@ pub fn optimize_region_file(region_file_path: &PathBuf, pb: &ProgressBar) -> std
             }
 
             if region.is_empty() {
-                pb.println(format!("[{:<15}] Deleted empty region", region_file_name));
-                std::fs::remove_file(region_file_path)?;
-                return Ok(());
+                result.deleted_regions += 1;
+                if write {
+                    std::fs::remove_file(region_file_path)?;
+                }
+                return Ok(result);
             }
 
             let deleted_chunk_count = chunks_to_delete.len();
             if deleted_chunk_count > 0 {
-                pb.println(format!(
-                    "[{:<15}] {} chunks deleted",
-                    region_file_name,
-                    chunks_to_delete.len()
-                ));
+                result.deleted_chunks += deleted_chunk_count;
             }
 
-            let bytes = region.to_bytes();
-            std::fs::write(region_file_path, bytes)?;
+            if write {
+                let bytes = region.to_bytes();
+                std::fs::write(region_file_path, bytes)?;
+            }
         }
-        Err(err) => {
-            pb.println(format!("[{:<15}] {}", region_file_name, err));
-            std::fs::remove_file(region_file_path)?;
+        Err(_) => {
+            result.deleted_regions += 1;
+            if write {
+                std::fs::remove_file(region_file_path)?;
+            }
         }
     }
 
-    Ok(())
+    Ok(result)
 }
